requirements.txt
fastapi
uvicorn
torch
pillow
streamlit
streamlit-drawable-canvas
requests
torchvision